more type-fixes to keep MSVC happy

+++ libambix/src/adaptor_acn.c
@@ -34,13 +34,13 @@ int _matrix_sid2acn(float32_t*data, uint32_t count) {
     uint32_t maxindex=ambix_order2channels(o)-offset;
 
     int32_t incr=2;
-    int32_t index;
+    uint32_t index;
 
     for(index=1; index<maxindex; index+=2) {
-      *datap++=index+offset;
+      *datap++=(float32_t)(index+offset);
     }
     for(index=maxindex-1; index>=0; index-=2) {
-      *datap++=index+offset;
+      *datap++=(float32_t)(index+offset);
     }
   }
   return 1;
